pyproject.toml
[tool.poetry]
name = "textese"
version = "0.1.0"
description = "Suite of tools for easily prep documents, flatten pdf's etc."
authors = ["Sergey Khalil <me@sergeyykhalil.com>"]
readme = "README.md"
packages = [{include = "textese"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.0"
pdf2image = "^1.16.3"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
flatten = "textese.flatten:main"
